Moving the code with the class to igrf_version.py
Rearranging __init__.py and moving the code with the class to igrf_version.py
Adding some changes in README.md
